@@ -10,21 +10,21 @@ def django_client(client) -> Client:
     return client
 
 
-@pytest.fixture
+@pytest.fixture()
 def user(django_user_model):
     return baker.make(django_user_model, is_staff=False)
 
 
-@pytest.fixture
+@pytest.fixture()
 def other_user(django_user_model):
     return baker.make(django_user_model, is_staff=False)
 
 
-@pytest.fixture
+@pytest.fixture()
 def admin(django_user_model):
     return baker.make(django_user_model, is_staff=True)
 
 
-@pytest.fixture
+@pytest.fixture()
 def superuser(django_user_model):
     return baker.make(django_user_model, is_staff=True, is_superuser=True)
